+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2019-11-27
+
 ### Added
 
 - New [`deep_update/3`][deep_update-3] function
@@ -36,7 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[unreleased]: https://github.com/eproxus/mapz/compare/v1.0.0...HEAD
+[unreleased]: https://github.com/eproxus/mapz/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/eproxus/mapz/compare/v1.0.0...v2.0.0
 [1.0.0]: https://github.com/eproxus/mapz/compare/v0.3.0...v1.0.0
 [0.3.0]: https://github.com/eproxus/mapz/releases/tag/v0.3.0